Setup lerna, move template files into component dir
