+++ lerna.json
@@ -1,0 +1,5 @@
+{
+  "packages": ["packages/*"],
+  "version": "independent",
+  "npmClientArgs": ["--legacy-peer-deps"]
+}
