+++ packages/annotation-toolkit/package.json
@@ -22,6 +22,7 @@
     "@blueprintjs/icons": "^3.26.0",
     "@typescript-eslint/eslint-plugin": "2.x",
     "@typescript-eslint/parser": "2.x",
+    "babel-eslint": "10.x",
     "babel-loader": "^8.1.0",
     "css-loader": "^3.5.1",
     "eslint": "6.x",
@@ -31,9 +32,11 @@
     "eslint-plugin-jsx-a11y": "6.x",
     "eslint-plugin-react": "7.x",
     "eslint-plugin-react-hooks": "2.x",
+    "file-loader": "^6.2.0",
     "global-context-store": "^1.0.1",
     "react": "16.13.1",
     "react-dom": "16.13.1",
+    "react-player": "^2.9.0",
     "style-loader": "^1.1.3",
     "url-loader": "^4.1.0",
     "webpack": "^4.42.1",
@@ -45,10 +48,5 @@
     "global-context-store": "^1.0.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1"
-  },
-  "dependencies": {
-    "babel-eslint": "10.x",
-    "file-loader": "^6.2.0",
-    "react-player": "^2.9.0"
   }
 }
